requirements.txt
google-endpoints==4.7.0
google-endpoints-api-management==1.11.0
validate_email==1.3